Fleshing out
Much more fleshed out.  Updates to all files.  Networks are composed of
Layers, Layers of Units.  The edges connecting the units of two
respective layers are represented as Matrices  At this stage, one can
construct Matrices, Layers, Networks, and Units.  The only piece yet to
be implemented is that which is most difficult: the iterative learning.
 *That* will be coming soon.