Builds
Changed all the instances of Matrix, usually replaced by arma::mat.
In a few cases, I replaced Matrix with vec, to reinforce the idea that
the matrix being passed is nothing more than a column vector.

It compiles, but I haven’t done debug tests yet to make sure that the
results are consistent with the original source.  That will be the next
thing to do.

**I want to verify, particularly, that the method “kron” used in the
Machine class (as a replacement for the inner product Matrix
constructor) behaves correctly.  I am almost certain that the way it’s
written is incorrect.  Namely, I’m taking two column vectors and taking
their IP.  That’s not going to work, I need to take one column vector
and one row vector.  I’m not sure how to control for that, but we’ll
see.