Minor Changes -- Huge Redesign Imminent
Layer
Changed some methods to satisfy the input of iron in the adjust weights
method.  It requires one column vector and one row vector to work
properly.

Machine
Changed the constructor, so that the machine is constructed how it’s
expected.  Armadillo uses column-ordered matrices, which was different
from my vector implementations.  I modified the constructor to match
Armadillo’s syntax.
Further, I modified some of the matrices to be row vectors, because
matrices were unnecessary.

I might need to make the same semantic changes I made in the Machine
constructor elsewhere, like in Network perhaps.  It remains to be seen.
 I still haven’t tested the output to see if it’s consistent with
pre-armadillo RBM.  That’s next on the list.

*    *    *

I added an omnigraffle sketch that represents a very ambitious
abstraction of this project, one that makes plenty of sense to me.
It’s a step in the direction of making this a proper library.  After
convincing myself that Armadillo is fully and properly integrated, I
will merge this branch with master and create a new branch entitled
RBM2.0.  That’ll be fun.  I’m going to flesh out the redesign with some
more detailed documentation before then, probably.

In the meantime, I made a makefile for compiling the project as-is,
with a main file that does almost nothing.  It was more a Proof of
Concept than a useful tool.  I’d never messed around with makefiles
before, so this was a good learning experience.

I’m even more excited now about the prospects of this project.