Formal Removal of the Matrix Classes
They aren’t used anymore, so… here we are!

Next up, testing the accuracy of this project vs the current main
branch project.  I want to make sure that Matrix is functionally
equivalent to arms::mat within the context of this project.

I’m basically going to generate some testing code that iterates through
the checklist of equivalencies outlined in this project’s readme.md.