Reverted Prev Changes; Did Testing; Squashed Bugs
There was a mistake in the Machine instructor, where the weights matrix
was instantiated in the transpose.  This is why I had missed needing to
transpose the result of the kron operation earlier.  I reverted the
changes I had naïvely made, having rectified the situation.

I tested the armadillo functionality against my Matrix class’s
functionality.  They are now functionally equivalent, with the
important caveat that I have to append “.t()” to every kron() call.
This is annoying, but it is what it is.  I’m now ready to merge into
Master, remove Matrix, and begin with the overhaul.